Implement inter-wiki link support
Symlinks are now treated as inter-wiki links, and followed to their conclusion.
Note that this changes the "current wiki" to be the wiki containing the target,
so all links there will be interpreted in that context. This may be initially
unintuitive, but the alternative (interpreting links in different namespaces based
on how one navigated to this entry) makes even less sense. </essay>